Add files via upload
Spark is a powerful, general purpose tool for working with Big Data. Spark transparently handles the distribution of computing tasks across a cluster. This means that operations are fast, but it also allows you to focus on the analysis rather than worry about technical details. In the document it is demonstrated  how to get data into Spark and then delve into the three fundamental Spark Machine Learning algorithms: Linear Regression, Logistic Regression/Classifiers, and creating pipelines. Along the way I have analysed a large dataset of flight delays and spam text messages. The codes in the documents can be used in any machine learning project with a little manupulation.